refactor(highlight): extract highlighting into its own crate
`okena-files` owned the syntect setup — the shared `SyntaxSet`, the theme
pair, the per-line span builder — and `okena-views-git` reached into it for
the diff viewer. That works while only viewers need colour, but the markdown
renderer needs it too, and `okena-files` already depends on `okena-markdown`,
so the dependency cannot run that way.

`okena-highlight` now sits below both. `syntax.rs` and `markdown_highlight.rs`
move across unchanged; `build_styled_text_with_backgrounds` moves out of
`code_view` for the same reason. `okena-files` re-exports all three, so every
`okena_files::syntax::…` and `code_view::build_styled_text_with_backgrounds`
import keeps working. Loading the `SyntaxSet` twice would have cost megabytes
for nothing, which is what a second copy in `okena-markdown` would have meant.

The name says highlighting rather than syntax because #188 is adding an
`okena-syntax` for something else — tree-sitter facts about code structure,
no colours involved.

`highlight_code_block` is the entry point a fenced block needs: it resolves a
language token rather than a file path, and loads the `SyntaxSet` itself since
the markdown renderer holds no file. It returns nothing for a fence with no
language or one syntect cannot place — plain-text highlighting would repaint
such a block in the syntax theme's foreground, and a block with nothing to
colour should keep the document's own text colour. Tabs are left alone there:
the caller maps character offsets onto the spans, so expanding a tab to four
spaces would shift every selection on the line. The file path keeps expanding
them, as it always has.

`syntax_for_lang` in `markdown_highlight` was the same mapping as the new
resolver and is now that resolver.
