+++ crates/okena-files/Cargo.toml
@@ -11,11 +11,8 @@ gpui = [
     "dep:gpui-component",
     "dep:okena-ui",
     "dep:okena-markdown",
+    "dep:okena-highlight",
     "dep:syntect",
-    "dep:two-face",
-    "dep:tree-sitter",
-    "dep:tree-sitter-md",
-    "dep:streaming-iterator",
     "dep:usvg",
     "dep:ttf-parser",
     "dep:image",
@@ -25,20 +22,15 @@ gpui = [
 okena-core = { path = "../okena-core" }
 okena-transport = { path = "../okena-transport", features = ["cancellable-http"] }
 okena-markdown = { path = "../okena-markdown", optional = true }
+okena-highlight = { path = "../okena-highlight", optional = true }
 okena-ui = { path = "../okena-ui", optional = true }
 
 gpui = { git = "https://github.com/zed-industries/zed", package = "gpui", optional = true }
 gpui-component = { git = "https://github.com/longbridge/gpui-component", package = "gpui-component", optional = true }
 
+# The highlighting itself lives in okena-highlight; syntect stays here for the
+# `SyntaxSet` the viewers hold and pass back in.
 syntect = { version = "5", features = ["default-fancy"], optional = true }
-two-face = { version = "0.5", default-features = false, features = ["syntect-fancy"], optional = true }
-
-# Markdown highlighting: tree-sitter is ~20x faster than syntect's Markdown
-# grammar on large files. tree-sitter (0.26) and streaming-iterator are already
-# in the workspace via gpui-component, so this only adds the grammar crate.
-tree-sitter = { version = "0.26", optional = true }
-tree-sitter-md = { version = "0.5", features = ["parser"], optional = true }
-streaming-iterator = { version = "0.1", optional = true }
 
 grep-searcher = "0.1"
 grep-regex = "0.1"
